Add dataset module and refactor pipeline
